hids_bot/requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
psutil>=5.9.0
python-telegram-bot>=13.7
SQLite3>=0.0.0
requests>=2.25.1 